requirements.txt
aiohttp>=3.13.3
amcrest==1.9.8
paho-mqtt==2.1.0
aiomqtt==2.1.0
backoff==2.2.1
coloredlogs==15.0.1
flvlib3@https://github.com/zkonge/flvlib3/archive/master.zip
hikvisionapi==0.3.2
packaging==24.1
Pillow>=11.3.0
uiprotect==10.2.2
reolinkapi==0.1.5
websockets>=14.0
xmltodict==0.13.0
pytapo==3.3.21
httpx~=0.28.0
pyjwt[crypto]>=2.12.1
